feat: Favorites icon is marked if movie is in favorites database.
The detail activity observes if there's been a change in favorites
database. If there is and the current movie id matches one in
the database, we know the movie is a favorite and mark it with the
star icon. The icon is also still displayed if screen is rotated.

Added function to set movies to favorites collection in
adapter.