feat: Change image sizes if viewing app on a tablet.
The image size strings for poster and backdrop URLS
were changed to static and set if the device is
detected to have a large screen size.
The app only checks the screen size when first started.
If the app is run on a phone or similarly sized device, the
image sizes remain default.