feat: Get complete YouTube URLs of a movie in its details activity.
Moved fetching video json from network into an asynctask.
Featurettes are excluded from the resulting list of video URLs
because they are too short and not satisfying to watch.

The video fetch task is executed in the details activity with movie id to
prevent long load time when the app first starts or is reloaded,
for example when main screen is rotated.