feat: Display user reviews and hide if there aren't any.
All of the first page of reviews are fetched and displayed
in a single TextView. If there are no reviews available, the
reviews label is set to View.Gone and the reviews TextView is
set to View.Invisible. A margin was added to the bottom of the
TextView to make screen look a little nicer.